learn_mqtt_kafka/kafka_mqtt_python/requirements.txt
paho-mqtt
kafka-python